# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "SponTan API",
    "description" : "Api Documentation",
    "termsOfService" : "urn:tos",
    "contact" : { },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/danielloza1108/SponTan/1.0",
    "description" : "SwaggerHub API Auto Mocking"
  }, {
    "url" : "http://localhost:8080",
    "description" : "Inferred Url"
  } ],
...
  "paths" : {
...
    "/api/event/add" : {
      "post" : {
        "tags" : [ "event-controller" ],
        "summary" : "addEvent",
        "operationId" : "addEventUsingPOST",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EventDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "x-content-type" : "*/*"
                }
              }
            }
          },
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/user/add" : {
      "post" : {
        "tags" : [ "user-controller" ],
        "summary" : "addUser",
        "operationId" : "addUserUsingPOST",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserDTO"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "x-content-type" : "*/*"
                }
              }
            }
          },
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/user/addFriend" : {
      "post" : {
        "tags" : [ "user-controller" ],
        "summary" : "addFriend",
        "operationId" : "addFriendUsingPOST",
        "requestBody" : {
          "description" : "Add friend to the user.",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/body"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "x-content-type" : "*/*"
                }
              }
            }
          },
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/user/edit/password" : {
      "post" : {
        "tags" : [ "user-controller" ],
        "summary" : "editPassword",
        "operationId" : "editPasswordUsingPOST",
        "requestBody" : {
          "description" : "Edit user's password.",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/body_1"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "x-content-type" : "*/*"
                }
              }
            }
          },
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/user/get/{email}" : {
      "get" : {
        "tags" : [ "user-controller" ],
        "summary" : "getUser",
        "operationId" : "getUserUsingGET",
        "parameters" : [ {
          "name" : "email",
          "in" : "path",
          "description" : "email",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/user/getFriendsId/{email}" : {
      "get" : {
        "tags" : [ "user-controller" ],
        "summary" : "getFriends",
        "operationId" : "getFriendsUsingGET",
        "parameters" : [ {
          "name" : "email",
          "in" : "path",
          "description" : "email",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "integer",
                    "format" : "int64"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/user/getUserById/{id}" : {
      "get" : {
        "tags" : [ "user-controller" ],
        "summary" : "getUserById",
        "operationId" : "getUserByIdUsingGET",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "EventDTO" : {
        "title" : "EventDTO",
        "type" : "object",
        "properties" : {
          "durationOfTheEvent" : {
            "type" : "string",
            "format" : "date-time"
          },
          "eventPlace" : {
            "type" : "string"
          },
          "eventStart" : {
            "type" : "string",
            "format" : "date-time"
          },
          "name" : {
            "type" : "string"
          },
          "quantityOfPlayers" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "UserDTO" : {
        "title" : "UserDTO",
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "example" : {
          "name" : "name",
          "email" : "email"
        }
      },
...
      "body" : {
        "type" : "object",
        "properties" : {
          "userEmail" : {
            "type" : "string"
          },
          "friendEmail" : {
            "type" : "string"
          }
        }
      },
      "body_1" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          }
        }
      }
    }
  }
}